+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [6.4.0]
+
+- Add `--excel-file` option to support XLSX file format in addition of CSV files.
+- Add `script/excel-downloader-browser.py`, a temporary workaround script that downloads an XLSX file from sharepoint by using the user browser session 
+- drop support of ruby 3.1
+
 ## [6.3.0] - 2024-11-04
 
 ### Removed